requirements.txt
anthropic>=0.40.0
notion-client>=2.2.1
python-dotenv>=1.0.0
